Superclass NetworkManager to get ip and port from network
WifiAwareNetwork extends NetworkManager to provide the ip and port to RtspClient. In other words, RtspClient obtains the ip and port using an instance of NetworkManager, without knowing that the network being used is WFA.

Deleted some Google.Guice imports because I didn't commit the gradle (and I don't plan using it for now).