Send proof zip in ANNOUNCE dsescription
Turn the zip file into a byte array and then into a Base64 string to append it into ANNOUNCE description using a custimized header ""a=proof:file=".

The receiver will parse it using regex, get the byte array, decode it back into a byte array and store it in ReceiveSession instance

ProofManager class: A singleton to get the zip File instance, the download directory and file name. The File instance to send it, and mDownloadDir and mFileHash is used by the receiver to store the file.